requirements.txt
fastapi==0.95.0
uvicorn==0.21.1
python-dotenv==1.0.0
psycopg2-binary==2.9.6
pyjwt==2.7.0
python-multipart==0.0.6
